Use null comparisons to see if things exist
This fixes a bug when logging in to the real server.

The `x == null` idiom is stricter than treating things as booleans, since it doesn't include "falsy" values like 0. It is also looser than the `in` operator, since it includes `undefined` and `null`.

Since the storage layer and login server use `null ` to indicate missing values, this balance is important.