Move the plugin objects out of redux
Putting the plugins into redux was a mistake to begin with, but there were no pixies at the time, so it seemed like a reasonable option.

With this change, the `currency` and `exchange` modules each manage their own plugins. The `currency` module does put the `currencyInfo` structures into redux, since those are necessary for mantaining a list of active wallets.