Do not process source files through babel es2015 for react-native
On Android we hit a problem whereby es2015 babel plugin will create nested functions inside IF statements which causes an error in RN since "use strict" is on by default. This change creates a separate build for react native that is only processed for flow and spread operators. This fixes Android and tested to still work on iOS.