Fix scrypt timing routine
Refactor to make calculations of Nrp much clearer to understand. Effective algorithm is the same although now we start with 16384, 8, 1 as the basis benchmark since increasing r has little effect on timing for most phones.

Major fix is to not pass a blank string to scrypt as the password or salt. This makes scrypt finish too quickly, on the order of 1-2ms regardless of Nrp parameters. This would cause high parameter estimates which would see huge delays when real passwords and salts are entered