Throttle calls to onTransactionsChanged and onNewTransactions to max every 5 seconds
This prevents the GUI from getting hammered with updates due to initial login which then cause calls to getTransactions over and over which are right now expensive due to the need to decrypt data off disk.

Change unit tests to wait after the first update as onTransactionsChanged and onNewTransactions may not be called synchronously with the update but rather after a setTimeout launched in the future.

Add unit tests which make sure back to back update calls get batched into one callback.