Un-hack the PIN login
Some accounts store the PIN in plain text. This makes it possible to toggle PIN login on or off without re-entering the PIN each time. Legacy PIN setups don't have this, so the user needs to enter their PIN if they toggle PIN login on and off.

Before, the PIN login method would try to capture the PIN in memory, emulating a new-style setup. Unfortunately, this creates an inconsistent experience with fingerprint login, password login, and so forth. This commit removes the hack, since it's incompatible with some changes coming in future commits.

A proper solution would notice the broken PIN and actually upgrade the account on the server, creating a good experience for future logins, regardless of the method used. Unfortunately, implementing that proper solution is too difficult at the moment. A bunch of cleanup needs to happen first.