Allow users to pass 6-digit OTP codes directly
Before, we only allowed the user to type in their 16-character backup code, which is less secure with screen readers.

Also, we called this option `otp`, instead of `otpKey` like it should have been. This commit still allows the old practice of passing `otpKey` as `otp` for compatiblity, but that way is deprecated now.

The login UI still requires changes to actually expose this feature.