Move custom tokens to the currency config object
Changes made here need to sync between all wallets in the account. We do this using a pixie to watch for differences, and by passing the initial set of custom tokens to each newly-created engine.

Since the existing `addCustomToken` method is part of an individual wallet or currency engine, we add a new `getTokenId` method at the currency-plugin level. This isn't mandatory, but if it's not present, the tradeoffs are that we cannot add tokens if an engine isn't running, and that we may call `addCustomToken` more than once on the same engine.