Send 'transactionsChanged' when editing metadata
We have always tried to send this callback, but we were filtering the transaction out because it had no on-chain changes. Currency engines do not send metadata, so we can conclude that the core sent the transaction if we get metadata. If the core sent the transaction, we should skip the filtering.