Add EdgeTxFiat data
This allows fiat plugins to tag transactions with fiat exchange specific data similar to EdgeTxSwap. This is however placed in the `currencies` object so we can tag different outgoing currencies with different information. EdgeTxSwap should really have been architected the same way but we'll need a migration plan for that.