Fix saving tokens at wallet creation time
This was a race condition between loading the tokens file off disk and writing the tokens file at wallet creation time. We solve the race condition by skipping the `dispatch` if there is no file on disk. That way, when we create a brand-new wallet, the initial tokens always win the race.