Fix an error while changing the PIN
We used to have two variables with the same name, `children`, but existing in separate scopes. Unfortunately, the minifier mis-compiled this code, referring to the inner variable instead of the outer variable. To fix this, we just give the outer variable a different name, forcing the minifier to refer to the correct one.