Fix in-memory metadata stomping bug
We only need to setup a new metadata file in memory for the transaction
if there doesn't already exists a file for the transaction. We only need
to do this so that the state is present for other places in the core
which depend on this state being present and deal with transaction
state.

In a future refactor, we can remove the setup entirely so long as we
remove the dependency on this state always being present for
transactions.