Fix loginWithPassword duress mode edge-case
This bug is where login with password from the login server does not
persist duress mode for the account, but persists it for other accounts
when attempting to login using a key.

The expected behavior is that login with password or key while in duress
mode should always bring you to the duress mode account/stash regardless
of whether login was online or offline. This is because online and
offline should always behave the same.