Change conditions for syncNetwork pixies
The syncNetwork pixies should never operate while the engine is not
started. This is a fix.

Also, the periodic-task pixie should be avoided while the wallet is
subscribing to the change-server. The core should properly mark a wallet
as avoiding sync-server subscriptions if the change-server is inoperable
or not supporting the wallet.