Show original error if pin-login is disabled
If the pin-login is disabled for the second try within loginWithPIN,
then we should show the original error message.

This is important for when you disable duress mode that we show the
invalid pin for the main account and also show the timeout message
instead of “pin not enabled” error.