Block pin-login on max wait between duress and main stashes
The issue is that there is a very easy way to get into a state where
the main account is in pin-waiting but the duress account is not. This
change stores the maximum wait time between the two login stashes and
blocks pin-login using this wait time. This prevents the user from ever
by-passing the wait time, leaking the duress mode feature.