Fix transpilation bug
Targeting ES5 transpilation causes a bug with looping over iterables.
The transpilation assumes all loops are over arrays, and so it loops
using `.length` properties. We cannot loop over iterables, therefore we
must convert the iterable to an array before looping.

This fixes a major bug with syncing caused by transpilation of the
source.