Add needsSync to EdgeEngineSyncNetworkOptions
This allows the core to differentiate between two events for the engine:

1. The engine is subscribed and needs to sync
2. The engine is subscribed and does not need to sync.

The engine unsuccessfully attempted to differentiate these two states
before by comparing the checkpoint with its own checkpoint, but this
was not correct because the checkpoint is the same value for both
states, making it impossible for the engine to know which of the two
states the wallet is in.

The fix is to introduce a `synced` state for the subscription which
will result in a `needsSync: false` state when calling `syncNetwork`.
This will tell the engine to complete its sync ratio without network
syncing.